utils.py don't remove trailing whitespace when parse 'show info' output
It can happen that the value of a field to be empty and the strip
causes the field name to not be stored in the dict.
Field description it is empty by default, the rest of the fields are
very unlikely to be empty as they are filled in by HAProxy.