Performance improvements due to the way we interact with stats socket
Up until now we use 'show stat' command for dumping all objects.
This patch changes this behavior by setting which type of objects to dump
but also which particular object.

HAProxy provides a way to filter the object to dump, from the doc:
show stat [<iid> <type> <sid>]
  Dump statistics in the CSV format. By passing <id>, <type> and <sid>, it is
  possible to dump only selected items :
    - <iid> is a proxy ID, -1 to dump everything
    - <type> selects the type of dumpable objects : 1 for frontends, 2 for
       backends, 4 for servers, -1 for everything. These values can be ORed,
       for example:
          1 + 2     = 3   -> frontend + backend.
          1 + 2 + 4 = 7   -> frontend + backend + server.
    - <sid> is a server ID, -1 to dump everything from the selected proxy.

Thats means we don't need to dump everything if we only interested in knowing
backend information.

This change of behavior improves a lot the performance of the library as we
only need to parse a very small portion of the data set.

This change will allow people with hundreds frontends/backends and thousands
of servers to get vital statistics(requests) very fast

Here is an example of the performance improvement using:
% grep '  server ' /etc/haproxy/haproxy.cfg|wc -l
10016
% grep '^backend' /etc/haproxy/haproxy.cfg|wc -l
105
% grep '^frontend' /etc/haproxy/haproxy.cfg|wc -l
105

We use github/unixsurfer/haproxytool for retrieving requests processed by
each frontend and measure the WLC using time cmd:

/usr/bin/time -f '%e' haproxytool -D /run/haproxy frontend -r >/dev/null;done

We run this 5 times, WLC time is in seconds:
     old  new code
1: 71.36 0.63
2: 73.16 0.62
3: 71.69 0.64
4: 71.45 0.78
5: 71.15 0.60
Avg:  72 0.65

We get ~82 times faster WLC!

How do we do that?

The important bit is to pass to internal objects(_Frontend, _Backend and
_Server) the unique ID they have during the instantiation. Any subsequent
'show stat' cmd will use the appropriate iid, obj_type and sid numbers.

We have to make sure that we always sent the correct IDs as a ID may be changed
after an addition or/and removal of frontend/backend/server sections in HAProxy
configuration.
HAProxy has to be reloaded in order for the new configuration to be applied, thus
we catch the change of IDs by checking if a process is reloaded before we send
the IDs to stats sockets.
The current detection works by checking if the creation time of the process
has changed during the life time of the execution of our code. In more details:

1) store pid and creation time during instantiation of _HAProxyProcess
2) reloaded function of _HAProxyProcess compares the creation time of
current process with the creation time during instantiation
3) process is reloead if those 2 times are different