connected_socket() perform a sanity on the date returned
The initial version of connected_socket was not checking data
returned from the socket and a result any random application
which listens to the socket file could be identified as a HAProxy.

We change the behavior to build a dictionary structure out of the
returned data and check if 'Name' key is present.