internal.py: remove empty string from data returned from socket
HAProxy always send an empty string at the end we remove it as
it adds noise for things like ACL/MAP and etc. We only do that
when we get more than 2 lines, which only happens when we ask
for ACL/MAP/etc and not for giving cmds such as disable/enable
server.