simplify the way we send commands to socket
We used to have one function(send_command) which was sending a command
to haproxy socket and was returning what socket returned to us.
Later we added a wrapper function(run_command) to filter the output, either
to return 1st line or the whole output. This layout adds extra code without any
real benefit. We just add full_output parameter to send_command to support the
same behavior and rename send_command to command.