Fix broken design in converter function
converter() was very badly designed and empty strings were slipped
through which caused bugs to other functions.

We now are more strict and raise ValueError if input value can't be
safely converted to integer.

Empty string or a sequence of spaces is converted to 0, it may not be
the right choice but it makes sense for our use-case where HAProxy
returns empty values for metrics.