Improve the way we internally use values for metrics
e8862894 and 6a714f670 commits revealed an ambiguous implementation on
how we use the values returned from HAProxy.

HAProxy may return for a certain metrics either a number or zero or
string or an empty string.  The functions converter and calculate
weren't properly designed to work with metrics which could have these
types of values.

converter() now tries to convert input to an integer or to a string if
conversion to number fails. If input value is an empty string returns
None. It also returns None if input is a data structure or object.

calculate() now handles properly empty lists and returns None in this
case.  Calls to round are removed as they aren't necessary.
The caller of calculate() has to filter out None values before they
passed in.

Converter() was called from metric() methods of internal objects
(_Frontend, _Backend and _Server), but now it being called from the
equivalent public objects where we also do the calculation.