Implement a proper retry logic for socket failures
This commit changes the retry logic we have in place for handling
connection failures on UNIX sockets(stats socket of HAProxy).

Possible values of retry parameter and their meaning:

- None => don't retry
- 0 => retry indefinitely
- 1..N => times to retry

Python doesn't provide a do while loop and @charlesnagy helped on building
a more pythonic way implementation.

This fixes also a bug where if retry was 0 we never tried to connect.