Return False when a file isn't a valid stats socket
connected_socket raises exceptions when the file isn't a valid HAProxy
stats socket which causes troubles in the way we use this function.

HAProxy class scans a directory for files and uses connected_socket()
to determine if files within the directory are valid HAProxy stats
sockets. If an exception raised by connected_socket then it propagates
it all the way up to the caller.

When HAProxy class scans the directory it only cares for valid files
and it doesn't need to care about invalid files. Thus, this commit
changes connected_socket to return False rather raising exceptions.