Don't check if ACL/MAP is a file
We get the content of MAP or ACL by either using an ID or the file
where entries are stored. haproxy expects "#" before ID, thus we must
detect if input is a file or a number. If the input isn't a number we
check if it is a valid file with with `os.path.isfile` and that check
causes problem when the file isn't available to the caller for various
reasons. In this issue https://github.com/unixsurfer/haproxytool/issues/5
for haproxytool the user runs haproxy inside a container and map files
are not accessible from the host.

In order to address this issue we skip the check if it is a valid file
and we simply send to haproxy what we receive. If input isn't a valid
file we get an error from haproxy:

    Unknown ACL identifier. Please use #<id> or <file>.