Try to smarter when we return address/port
haproxy returns the assigned address and port of a server as <address:port>
and when we compare values across all processes we may detect
inconsistencies either on port or address or on both.

When we want to report only one of them, let's say the address, we don't
want to raise IncosistentData exception in case only the port is
different per process.

Thus, we check if the inconsistency is for the property we want to report,
either address or port, and if it isn't then we simply report what we have.
In case the inconsistency is for the property we want to report then we
raise IncosistentData exception.